Fix compile warnnings
# message auto-generated for no-merge-commit merge:
merge fix into main

Fix compile warnnings

Created-by: Ooohui
Commit-by: Ooohui
Merged-by: liwei3013
Description: Fix compile warnnings

See merge request: openHiTLS/openhitls!151

+++ bsl/asn1/src/bsl_asn1_local.h
@@ -32,13 +32,13 @@ extern "C" {
 #define BSL_ASN1_CLASS_MASK            0xC0
 
 typedef struct _ASN1_AnyOrChoiceParam {
-    int32_t idx;
+    uint32_t idx;
     void *previousAsnOrTag;
     BSL_ASN1_DecTemplCallBack tagCb;
 } BSL_ASN1_AnyOrChoiceParam;
 
 typedef struct _BSL_ASN1_EncodeItem {
-    uint64_t asnOctetNum; // tag + len + content
+    uint32_t asnOctetNum; // tag + len + content
     BSL_ASN1_Buffer *asn;
     uint8_t tag;
     uint8_t depth;
